Refactor print_frequencies
