+++ lookout/style/typos/tests/test_corrector_utils.py
@@ -1,6 +1,7 @@
 from os.path import join
 import pathlib
 import pickle
+import tempfile
 import unittest
 
 import numpy
@@ -8,11 +9,39 @@ import pandas
 from pandas.util.testing import assert_frame_equal
 
 from lookout.style.typos.utils import (
-    add_context_info, Columns, flatten_df_by_column,
+    add_context_info, Columns, filter_splits, flatten_df_by_column, print_frequencies,
     rank_candidates, read_frequencies, read_vocabulary, suggestions_to_df, suggestions_to_flat_df)
 
 
 TEST_DATA_PATH = str(pathlib.Path(__file__).parent)
+
+
+class PreprocessingTest(unittest.TestCase):
+    def test_filter_splits(self):
+        data = pandas.DataFrame([["get value", "get"],
+                                 ["get value", "value"],
+                                 ["gut", "gut"],
+                                 ["put tok", "put"],
+                                 ["put tok", "tok"],
+                                 ["put", "put"]], columns=[Columns.Split, Columns.Token])
+        vocabulary = {"get", "value", "put"}
+        result = pandas.DataFrame([["get value", "get"],
+                                   ["get value", "value"],
+                                   ["put", "put"]], columns=[Columns.Split, Columns.Token],
+                                  index=[0, 1, 5])
+        assert_frame_equal(filter_splits(data, vocabulary), result)
+
+    def test_print_frequencies(self):
+        vocabulary = {"get": 23, "set": 17}
+        with tempfile.NamedTemporaryFile() as vocabulary_file:
+            print_frequencies(vocabulary, vocabulary_file.name)
+            with open(vocabulary_file.name) as f:
+                lines = []
+                for line in f:
+                    lines.append(line.strip())
+            self.assertEqual(len(lines), 2)
+            self.assertIn("get,23", lines)
+            self.assertIn("set,17", lines)
 
 
 class ReadDataTest(unittest.TestCase):
